d2m greedy optimizer integration draft
