+++ lib/Dialect/TTNN/Transforms/OptimizerPasses/Optimizer.cpp
@@ -23,6 +23,7 @@
 #include "ttmlir/Dialect/TTNN/Interfaces/TTNNTensorSpecInterface.h"
 #include "ttmlir/Dialect/TTNN/Pipelines/TTNNPipelines.h"
 #include "ttmlir/Dialect/TTNN/Transforms/Passes.h"
+#include "ttmlir/Dialect/TTNN/Utils/D2MOptimizerUtils.h"
 #include "ttmlir/Dialect/TTNN/Utils/PassOverrides.h"
 #include "ttmlir/Dialect/TTNN/Utils/Utils.h"
 #include "ttmlir/FunctionTypes.h"
@@ -65,120 +66,6 @@ TTNNOptimizerOptions::TTNNOptimizerOptions(
 namespace {
 
 #ifdef TTMLIR_ENABLE_OPMODEL
-
-/// Applies the chosen layout to a D2MSubgraphOp: result type(s), output
-/// buffer(s) (Empty op), and the referenced D2M subgraph function.
-void applyChosenLayoutToD2MSubgraphOp(D2MSubgraphOp dispatchOp,
-                                      RankedTensorType newTensorType,
-                                      TTNNLayoutAttr layoutAttr,
-                                      ttcore::GridAttr deviceGrid) {
-  assert(dispatchOp.getNumResults() <= 1 &&
-         "D2MSubgraphOp with multiple results not yet supported");
-
-  for (unsigned i = 0; i < dispatchOp.getNumResults(); ++i) {
-    dispatchOp.getResult(i).setType(newTensorType);
-  }
-
-  for (Value output : dispatchOp.getOutputs()) {
-    if (EmptyOp emptyOp = output.getDefiningOp<EmptyOp>()) {
-      emptyOp.getResult().setType(newTensorType);
-      BufferType bufferType = layoutAttr.getBufferType();
-      TensorMemoryLayoutAttr tensorMemoryLayoutAttr = layoutAttr.getMemLayout();
-      emptyOp.setDtype(layoutAttr.getDataType());
-      if (layoutAttr.isTiled()) {
-        emptyOp.setLayout(ttnn::Layout::Tile);
-      } else {
-        emptyOp.setLayout(ttnn::Layout::RowMajor);
-      }
-      emptyOp.setMemoryConfigAttr(ttnn::MemoryConfigAttr::get(
-          dispatchOp.getContext(), tensorMemoryLayoutAttr,
-          BufferTypeAttr::get(dispatchOp.getContext(), bufferType),
-          utils::createShardSpecIfNeeded(layoutAttr, deviceGrid)));
-    } else {
-      llvm::report_fatal_error(
-          "Expected EmptyOp for D2MSubgraphOp output buffer");
-    }
-  }
-
-  if (func::FuncOp mainFunc = dispatchOp.getD2MMainFunc()) {
-    Block &entryBlock = mainFunc.getBody().front();
-    unsigned argIdx = 0;
-    for (Value input : dispatchOp.getInputs()) {
-      if (argIdx < entryBlock.getNumArguments()) {
-        Type inputType = input.getType();
-        if (isa<RankedTensorType>(inputType)) {
-          entryBlock.getArgument(argIdx).setType(inputType);
-        }
-        ++argIdx;
-      }
-    }
-    // Insert a trailing to_layout to convert the last op's output to the
-    // chosen layout X instead of rewriting every body op's result type.
-    Block &block = mainFunc.getBody().front();
-    Operation *terminator = block.getTerminator();
-    if (func::ReturnOp returnOp = dyn_cast<func::ReturnOp>(terminator)) {
-      if (returnOp.getNumOperands() > 0) {
-        Value currentResultValue = returnOp.getOperand(0);
-        if (currentResultValue.getType() != newTensorType) {
-          OpBuilder builder(dispatchOp.getContext());
-          builder.setInsertionPoint(returnOp);
-          ttcore::DataTypeAttr dataType = ttcore::DataTypeAttr::get(
-              dispatchOp.getContext(), layoutAttr.getDataType());
-          LayoutAttr newLayout =
-              LayoutAttr::get(dispatchOp.getContext(), layoutAttr.getLayout());
-          MemoryConfigAttr memConfigAttr = MemoryConfigAttr::get(
-              dispatchOp.getContext(), layoutAttr.getMemLayout(),
-              BufferTypeAttr::get(dispatchOp.getContext(),
-                                  layoutAttr.getBufferType()),
-              utils::createShardSpecIfNeeded(layoutAttr, deviceGrid));
-          Location loc = mainFunc.getLoc();
-          ToLayoutOp toLayoutOp =
-              builder.create<ToLayoutOp>(loc, newTensorType, currentResultValue,
-                                         newLayout, dataType, memConfigAttr);
-          returnOp.setOperand(0, toLayoutOp.getResult());
-        }
-      }
-    }
-    SmallVector<Type> newInputTypes;
-    for (Value input : dispatchOp.getInputs()) {
-      newInputTypes.push_back(input.getType());
-    }
-    SmallVector<Type> newResultTypes(dispatchOp.getNumResults(), newTensorType);
-    mainFunc.setType(FunctionType::get(dispatchOp.getContext(), newInputTypes,
-                                       newResultTypes));
-  }
-}
-
-/// Sync the D2M subgraph function's argument and function types to the
-/// dispatch op's current input types (e.g. after spill, first operand may be
-/// DRAM).
-void syncD2MFuncTypesToDispatchInputs(D2MSubgraphOp dispatchOp) {
-  func::FuncOp mainFunc = dispatchOp.getD2MMainFunc();
-  if (!mainFunc) {
-    return;
-  }
-  Block &entryBlock = mainFunc.getBody().front();
-  unsigned argIdx = 0;
-  for (Value input : dispatchOp.getInputs()) {
-    if (argIdx < entryBlock.getNumArguments()) {
-      Type inputType = input.getType();
-      if (isa<RankedTensorType>(inputType)) {
-        entryBlock.getArgument(argIdx).setType(inputType);
-      }
-      ++argIdx;
-    }
-  }
-  SmallVector<Type> newInputTypes;
-  for (Value input : dispatchOp.getInputs()) {
-    newInputTypes.push_back(input.getType());
-  }
-  SmallVector<Type> newResultTypes;
-  for (Value result : dispatchOp.getResults()) {
-    newResultTypes.push_back(result.getType());
-  }
-  mainFunc.setType(FunctionType::get(dispatchOp.getContext(), newInputTypes,
-                                     newResultTypes));
-}
 
 #endif // TTMLIR_ENABLE_OPMODEL
 
@@ -566,8 +453,8 @@ public:
           // D2MSubgraphOp: apply chosen layout to result(s), output buffer(s),
           // and D2M subgraph.
           if (auto dispatchOp = dyn_cast<D2MSubgraphOp>(op)) {
-            applyChosenLayoutToD2MSubgraphOp(dispatchOp, newTensorType,
-                                             layoutAttr, deviceGrid);
+            d2m_optimizer_utils::applyChosenLayoutToD2MSubgraphOp(
+                dispatchOp, newTensorType, layoutAttr, deviceGrid);
             return;
           }
 
@@ -713,9 +600,7 @@ public:
 
       // Sync D2M subgraph function types to match dispatch op's current inputs
       // (e.g. after spill, first operand may be DRAM).
-      func->walk([&](D2MSubgraphOp dispatchOp) {
-        syncD2MFuncTypesToDispatchInputs(dispatchOp);
-      });
+      d2m_optimizer_utils::syncAllD2MFuncTypes(func);
 
       // Try finding ops that can be upgraded from DRAM to L1 interleaved
       // layout.
